Add files via upload
This is a Career Development Criteria Self Marketing Critical Thinking assignment. The attached project a summary of my Capstone project for the course of Business intelligence and Data Analytics. The project is about the CDC-Pfizer Covid-19 vaccine address the Pfizer COVID-19 efficacy, herd immunity in the U.S, and assess the goal of Whitehouse to vaccinate 70% of Adult Americans or to vaccinate 160 Million adulate Americans by July 04. to predict the number of people SAS studio is applied.